Moved testing to predict.py
